+++ product-lucra-gyp/ios-app/LucraSDKTarget/LucraSDKTarget.swift
@@ -1,0 +1,10 @@
+//
+//  File.swift
+//  
+//
+//  Created by Michael Schmidt on 7/12/23.
+//
+
+import Foundation
+
+// Just need an empty file for binary framework wrapper
