+++ product-lucra-gyp/ios-app/Package.swift
@@ -1,0 +1,48 @@
+// swift-tools-version: 5.8
+// The swift-tools-version declares the minimum version of Swift required to build this package.
+
+import PackageDescription
+
+let version = "3.2.0"
+
+let hostedPackageURL = "https://lucra-sdk.s3.amazonaws.com/ios/spm/\(version)"
+
+let package = Package(
+    name: "LucraSDK",
+    platforms: [.iOS(.v15)],
+    products: [
+        .library(name: "LucraSDK", targets: ["LucraSDK", "LucraCore"])
+    ],
+    dependencies: [
+        .package(url: "https://github.com/zendesk/support_sdk_ios", .upToNextMajor(from: "8.0.0")),
+        .package(url: "https://github.com/auth0/Auth0.swift.git", .upToNextMajor(from: "2.0.0"))
+    ],
+    targets: [
+        .target(
+            name: "LucraCore",
+            dependencies: [
+                "LucraSDK",
+                "MobileIntelligence",
+                "GeoComplySDK",
+                .product(name: "Auth0", package: "Auth0.swift"),
+                .product(name: "ZendeskSupportSDK", package: "support_sdk_ios")
+            ],
+            path: "LucraSDKTarget"
+        ),
+        .binaryTarget(
+            name: "LucraSDK",
+            url: "\(hostedPackageURL)/LucraSDK.xcframework.zip",
+            checksum: "f88792c4894e5ea709a8635edd838705f1281f8801d7368ed7ae4df4e6f59dcf"
+        ),
+        .binaryTarget(
+            name: "MobileIntelligence",
+            url: "\(hostedPackageURL)/MobileIntelligence.xcframework.zip",
+            checksum: "89533fab570a147ecfa924096b6358e3ec748ffda10f63bee8380c2a5fb25b7e"
+        ),
+        .binaryTarget(
+            name: "GeoComplySDK",
+            url: "\(hostedPackageURL)/GeoComplySDK.xcframework.zip",
+            checksum: "c805a9e25b2093f5bfe65edcdc2b2ef1c8dc1042feb799da92e4f82732c35550"
+        )
+    ]
+)
