+++ product-lucra-gyp/ios-app/RNGApp/Model/Stats.swift
@@ -1,0 +1,13 @@
+//
+//  Stats.swift
+//  RNGApp
+//
+//  Created by Wellison Pereira on 8/12/25.
+//
+
+import Foundation
+
+struct Stats: Codable {
+    var totalNumbersGenerated: Int
+    var bestNumber: Int
+}
