+++ product-lucra-gyp/ios-app/RNGApp/RNGApp.swift
@@ -1,9 +1,10 @@
 import SwiftUI
+import LucraSDK
 
 @main
 struct RNGApp: App {
     @StateObject private var session = SessionManager()
-
+    
     var body: some Scene {
         WindowGroup {
             MainView()
