+++ product-lucra-gyp/ios-app/RNGApp/View/Main/MainView.swift
@@ -1,7 +1,10 @@
+
 import SwiftUI
+import LucraSDK
 
 struct MainView: View {
     @EnvironmentObject var session: SessionManager
+    @EnvironmentObject var client: LucraClient
 
     var body: some View {
         NavigationStack {
